+++ docs/changelog.rst
@@ -36,6 +36,7 @@ The format is based on `Keep a Changelog`_, and this project adheres to `Semanti
 
 -  Incorrect handling of services and characteristics in the Python examples. *(Thanks Carl-CWX!)*
 -  Minor potential race condition in the safe callback.
+-  (Linux) Peripheral would still issue callbacks after deletion.
 -  (MacOS) Increased priority of the dispatch queue to prevent jitter in the incoming data.
 -  (Windows) Missing peripheral identifier data. *(Thanks eriklins!)*
 -  (Windows) Multiple initialization of the WinRT backend.